add p2p cmd
